add a uniq id to nohistory dialogs
if won't pop up the dialog on initial load (`nohistory`) the user MAY
refresh a url with the dialog id as the hash then a change of the
hash won't be recognized by the browser when the dialog comes up and
the back button will return to the referring page. So, when nohistory is
defined, we append a "unique" identifier to the hash.

Fixes #32