fix back button failure
Compare the initial hash with the current hash to ensure
that the browser doesn't move back past the hash from when
the dialog was instantiated.

Notably the current solution isn't particularly graceful in
cases where the dialog is intantiated late in the life of the page
(possibly after hashes have been added). We could record the initial
hash at load time for comparison.